+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # 2.3.4
+- Give all BWG Entities Eye Heights
+- Increase Oddion Hitbox Size
+- Decrease Pumpkin Warden Hitbox Size
+- Move Oddion to Entity Type Creature
+  -This Change should fix large amounts of Oddions spawning in the same areas 
+- Move ManOWar to Entity Type Water Creature
 
 # 2.3.3
 - Add ModMenu Support (Fabric)